+++ src/headless/index.js
@@ -8,6 +8,12 @@ import u from './utils/index.js';
 import converse from './shared/api/public.js';
 import log from './log.js';
 
+import BaseMessage from './shared/message.js';
+export { BaseMessage };
+
+import ModelWithMessages from './shared/model-with-messages.js';
+export { ModelWithMessages };
+
 // START: Removable components
 // ---------------------------
 // The following components may be removed if they're not needed.
@@ -21,9 +27,6 @@ import './plugins/chatboxes/index.js';
 import './plugins/disco/index.js'; // XEP-0030 Service discovery
 import './plugins/adhoc/index.js'; // XEP-0050 Ad Hoc Commands
 import './plugins/headlines/index.js'; // Support for headline messages
-
-import ModelWithMessages from './shared/model-with-messages.js';
-export { ModelWithMessages };
 
 // XEP-0313 Message Archive Management
 export { MAMPlaceholderMessage } from './plugins/mam/index.js';