+++ src/types/shared/chat/utils.d.ts
@@ -78,21 +78,11 @@ export function shortnamesToEmojis(str: string, options?: {
     unicode_only: boolean;
     add_title_wrapper: boolean;
 }): any[];
-/**
- * Called when the chat content is scrolled up or down.
- * We want to record when the user has scrolled away from
- * the bottom, so that we don't automatically scroll away
- * from what the user is reading when new messages are received.
- */
-export const markScrolled: import("lodash").DebouncedFunc<(ev: Event) => void>;
 export type EmojiMarkupOptions = {
     unicode_only?: boolean;
     add_title_wrapper?: boolean;
 };
 export type HeadingButtonAttributes = import("plugins/chatview/types").HeadingButtonAttributes;
-export type ChatView = import("../../plugins/chatview/chat.js").default;
-export type MUCView = import("../../plugins/muc-views/muc.js").default;
-export type MUCOccupantView = import("../../plugins/muc-views/occupant").default;
 export type Message = import("@converse/headless").Message;
 export type MUCMessage = import("@converse/headless").MUCMessage;
 export type Model = import("@converse/skeletor").Model;